Moved ratatui logic to another module
