+++ src/scraper.rs
@@ -1,12 +1,12 @@
+use crate::driver::*;
+use ntfy::{dispatcher, Error, Payload, Priority, Url};
+use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
 use std::fmt;
 use std::fs::{read_to_string, File};
 use std::io::Write;
-use ntfy::{dispatcher, Error, Payload, Priority, Url};
 use thirtyfour::error::WebDriverResult;
 use thirtyfour::{By, WebDriver};
-use serde::{Deserialize, Serialize};
-use crate::driver::*;
 
 #[derive(Clone)]
 pub(crate) struct Scraper {
@@ -32,7 +32,11 @@ pub(crate) struct ScraperConfig {
 }
 impl fmt::Display for ScraperConfig {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "({}, {}, {}, {})", self.name, self.url, self.base_url_to_prepend, self.house_link_css)
+        write!(
+            f,
+            "({}, {}, {}, {})",
+            self.name, self.url, self.base_url_to_prepend, self.house_link_css
+        )
     }
 }
 impl Scraper {
@@ -112,14 +116,14 @@ impl Scraper {
     }
 
     pub(crate) async fn run(&mut self) {
-        self.load_previous_session_file()
-            .expect("Unexpected error in reading the file");
+        // self.load_previous_session_file()
+        //     .expect("Unexpected error in reading the file");
         loop {
             //TODO: manage scraped_results fail
             let scraped_results = self.scrape().await.expect("TODO: panic message");
             self.detect(scraped_results).await;
             tokio::time::sleep(tokio::time::Duration::from_secs(120)).await;
-        } 
+        }
     }
 }
